% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagrangeMgmtDesign.R
\name{LagrangeMgmtDesign}
\alias{LagrangeMgmtDesign}
\alias{LagrangeMgmtDesign.ManageContext}
\title{Lagrange management design class builder}
\usage{
LagrangeMgmtDesign(
  context,
  divisions,
  establish_pr,
  f_obj,
  f_deriv,
  f_pos,
  alpha_unconstr,
  alpha_min,
  f_unit_eff,
  f_inv_unit_eff,
  budget = NULL,
  average_pr = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  search_alpha = FALSE,
  ...
)

\method{LagrangeMgmtDesign}{ManageContext}(
  context,
  divisions,
  establish_pr,
  f_obj,
  f_deriv,
  f_pos,
  alpha_unconstr,
  alpha_min,
  f_unit_eff,
  f_inv_unit_eff,
  budget = NULL,
  average_pr = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  search_alpha = FALSE,
  ...
)
}
\arguments{
\item{context}{A \code{ManageContext} or inherited class object representing
the context of a biosecurity management resource allocation design.}

\item{divisions}{A \code{ManageDivisions} or inherited class object
representing one or more divisions (management responses, strategies,
invasion pathways, invasive species, spatial locations, time, etc.) for
the management design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part (location,
category, etc.) specified by \code{divisions}. Values are assumed to be
relative when their maximum is greater than 1, or an attribute
\code{relative = TRUE} is attached to the parameter.}

\item{f_obj}{The objective function for calculating values to optimize
(minimize) in the form \code{function(x_alloc)}, where \code{x_alloc}
represents a candidate resource allocation at each division part
(location, category, etc.) specified by \code{divisions}.}

\item{f_deriv}{The derivative of the objective function \code{f_obj} in the
form \code{function(x_alloc)}, where \code{x_alloc} represents a candidate
resource allocation at each division part (location, category, etc.)
specified by \code{divisions}.}

\item{f_pos}{The pseudo-inverse of the derivative function \code{f_deriv} in
the form \code{function(alpha)}, where \code{alpha} represents the
marginal benefit value for a candidate resource allocation \code{x_alloc},
which is returned by the function for each division part (location,
category, etc.) specified by \code{divisions}.}

\item{alpha_unconstr}{The marginal benefit value to utilize when the search
for the optimal resource allocation is not constrained via a \code{budget}
or a desired overall effectiveness \code{overall_pr}.}

\item{alpha_min}{The minimum marginal benefit value to utilize when
searching for the optimal resource allocation.}

\item{f_unit_eff}{A function for calculating the unit (division part)
effectiveness, or probability of successful management of an incursion
when present. The function should be in the form \code{function(x_alloc)},
where \code{x_alloc} represents a candidate resource allocation at each
division part (location, category, etc.) specified by \code{divisions}.}

\item{f_inv_unit_eff}{A function for calculating the inverse of the unit
effectiveness, or candidate resource allocation. The function should be in
the form \code{function(unit_eff)}, where \code{unit_eff} represents the
unit effectiveness at each division part (location, category, etc.)
specified by \code{divisions}.}

\item{budget}{The cost budget or constraint for the resource allocation in
the management design. Default is \code{NULL}.}

\item{average_pr}{The desired (minimum) weighted average probability of
success (or effectiveness) of the management design (e.g. 0.95). The
weighted average is calculated using (relative) establishment probability
(\code{establish_pr}) values. Default is \code{NULL}.}

\item{overall_pr}{The desired (minimum) overall system-wide probability of
success (or effectiveness) of the management design (e.g. 0.95). Can only
be used when actual (not relative) establishment probability
(\code{establish_pr}) values are provided. Default is \code{NULL}.}

\item{min_alloc}{A vector of minimum permissible allocated management
resource quantities at each division part (location, category, etc.)
specified by \code{divisions}. Used to avoid impractically low allocation
quantities. Default is \code{NULL}.}

\item{search_alpha}{A logical indicator to search for the optimal resource
allocation, even when it is not constrained (via \code{budget} or
\code{overall_pr}), such as when fixed costs are present. Default is
\code{FALSE} indicates that search only occurs when constrained.}

\item{...}{Additional parameters.}
}
\value{
A \code{LagrangeMgmtDesign} class object (list) containing functions
for allocating resource costs:
\describe{
\item{\code{get_cost_allocation()}}{Get allocated management resources
(costs) via Lagrange-based method, utilizing savings, benefits, budget
constraints, and/or desired system-wide effectiveness level.}
}
}
\description{
Builds a generic class to represent management design functionality for
the effective allocation of management resources across one or more
divisions (parts, locations, categories, etc.) via Lagrange-based methods
for optimizing objective functions specified with management costs,
savings, benefits, and/or overall probability of management success (or
effectiveness).
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}

Hauser, C. E., & McCarthy, M. A. (2009). Streamlining 'search and
destroy': cost-effective surveillance for invasive species management.
\emph{Ecology Letters}, 12(7), 683–692.
\doi{10.1111/j.1461-0248.2009.01323.x}

Giljohann, K. M., Hauser, C. E., Williams, N. S. G., & Moore, J. L.
(2011). Optimizing invasive species control across space: willow invasion
management in the Australian Alps. \emph{Journal of Applied Ecology},
48(5), 1286–1294. \doi{10.1111/j.1365-2664.2011.02016.x}

McCarthy, M. A., Thompson, C. J., Hauser, C., Burgman, M. A., Possingham,
H. P., Moir, M. L., Tiensin, T., & Gilbert, M. (2010). Resource allocation
for efficient environmental management. \emph{Ecology Letters}, 13(10),
1280–1289. \doi{10.1111/j.1461-0248.2010.01522.x}

Moore, A. L., McCarthy, M. A., & Lecomte, N. (2016). Optimizing ecological
survey effort over space and time.
\emph{Methods in Ecology and Evolution}, 7(8), 891–899.
\doi{10.1111/2041-210X.12564}
}
