% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageContext.R
\name{ManageContext}
\alias{ManageContext}
\alias{ManageContext.default}
\title{Manage context class builder}
\usage{
ManageContext(
  species_names,
  species_types = c("pest", "weed", "disease"),
  resource_type = c("survey", "traps", "treatment", "removal", "mixed", "other"),
  management_purpose = c("delimitation", "containment", "eradication", "other"),
  threat_status = c("detected", "delimited", "contained", "eradicated", "other"),
  ...
)

\method{ManageContext}{default}(
  species_names,
  species_types = c("pest", "weed", "disease"),
  resource_type = c("survey", "traps", "treatment", "removal", "mixed", "other"),
  management_purpose = c("delimitation", "containment", "eradication", "other"),
  threat_status = c("detected", "delimited", "contained", "eradicated", "other"),
  ...
)
}
\arguments{
\item{species_names}{Vector of one or more invasive species (or genus)
names.}

\item{species_types}{Vector of types of invasive species. One or more of
\code{"pest"}, \code{"weed"}, or \code{"disease"}.}

\item{resource_type}{The type of management resource being allocated. One of
\code{"survey"}, \code{"traps"}, \code{"treatment"}, \code{"removal"},
\code{"mixed"}, or \code{"other"}.}

\item{management_purpose}{The purpose of the management. One of
\code{"delimitation"}, \code{"containment"}, \code{"eradication"}, or
\code{"other"}.}

\item{threat_status}{The status of the invasive species threat. One of
\code{"detected"}, \code{"delimited"}, \code{"contained"},
\code{"eradicated"}, or \code{"other"}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageContext} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_species_names()}}{Get the invasive species names.}
\item{\code{get_species_types()}}{Get the types of invasive species:
"pest", "weed", or "disease".}
\item{\code{get_resource_type()}}{Get the resource type.}
\item{\code{get_management_purpose()}}{Get the management purpose.}
\item{\code{get_threat_status()}}{Get the threat status.}
}
}
\description{
Builds a class to represent the context of a biosecurity management resource
allocation, including information about the invasive pest, weed or disease
species or genus being monitored, the type of management resources
allocated, the purpose of the management, and the status of the threat.
}
