% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageDetection.R
\name{ManageDetection}
\alias{ManageDetection}
\alias{ManageDetection.Region}
\title{Manage detection class builder}
\usage{
ManageDetection(
  region,
  population_model,
  surveillance,
  stages = NULL,
  schedule = NULL,
  ...
)

\method{ManageDetection}{Region}(
  region,
  population_model,
  surveillance,
  stages = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{surveillance}{A \code{bsdesign::SurveillanceDesign} or inherited class
object representing the distribution of surveillance resources and their
detection sensitivities.}

\item{stages}{Numeric vector of population stages (indices) to which
management detection are applied. Default is all stages (when set to
\code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps in which to apply
management detection. Default is all time steps (when set to \code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageDetection} class object (list) containing a function
for accessing attributes and applying simulated management detection:
\describe{
\item{\code{get_type()}}{Get the type of management action
("detection").}
\item{\code{get_label()}}{Get the management actions label used in
simulation results (i.e. "detected").}
\item{\code{get_surveillance()}}{Get the surveillance design class
object.}
\item{\code{get_stages()}}{Get the population stages to which management
detection are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which management detection are applied.}
\item{\code{apply(n)}}{Apply management detection to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, and return the
resulting population \code{n} along with attached attributes relating
to the newly applied detection/surveillance.}
}
}
\description{
Builds a class for simulating the application of detection or surveillance
resources applied in the management of an invasive species.
}
