% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageRemovals.R
\name{ManageRemovals}
\alias{ManageRemovals}
\alias{ManageRemovals.Region}
\title{Manage removals class builder}
\usage{
ManageRemovals(region, population_model, apply_stages = NULL, ...)

\method{ManageRemovals}{Region}(region, population_model, apply_stages = NULL, ...)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{apply_stages}{Numeric vector of population stages (indices) to which
management removals are applied. Default is all stages (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageRemovals} class object (list) containing a function
for accessing attributes and applying simulated management removals:
\describe{
\item{\code{get_type()}}{Get the type of management action ("removal").}
\item{\code{apply(n)}}{Apply management removals to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, and return the
resulting population \code{n} along with attached attributes relating
to the newly applied removals.}
}
}
\description{
Builds a class for simulating the application of management removals or
eradication.
}
