% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlDesign.R
\name{ControlDesign}
\alias{ControlDesign}
\alias{ControlDesign.ManageContext}
\title{Control design class builder}
\usage{
ControlDesign(
  context,
  divisions,
  dim_type = c("spatial", "species", "pathways", "user"),
  establish_pr,
  lambda = NULL,
  unit_pr = NULL,
  optimal = c("saving", "benefit", "successes", "effectiveness", "none"),
  alloc_unit = c("units", "hours", "traps", "treatments", "removals", "user"),
  cost_unit = c("$", "hours", "user"),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  average_pr = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_manage_pr = NULL,
  previous_control = NULL,
  class = character(),
  ...
)

\method{ControlDesign}{ManageContext}(
  context,
  divisions,
  dim_type = c("spatial", "species", "pathways", "user"),
  establish_pr,
  lambda = NULL,
  unit_pr = NULL,
  optimal = c("saving", "benefit", "successes", "effectiveness", "none"),
  alloc_unit = c("units", "hours", "traps", "treatments", "removals", "user"),
  cost_unit = c("$", "hours", "user"),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  average_pr = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_manage_pr = NULL,
  previous_control = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{ManageContext} or inherited class object representing
the context of a biosecurity management resource allocation design.}

\item{divisions}{A \code{ManageDivisions} or inherited class object
representing one or more divisions (invasion pathways, invasive species,
spatial locations, etc.) for the management design.}

\item{dim_type}{The type of dimension that the management resources are
allocated across. One of \code{"spatial"} locations (default), invasive
\code{"species"}, invasion \code{"pathways"}, or other user defined. The
dimensions should be appropriately configured via the \code{divisions}
object.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part specified by
\code{divisions}. Values are assumed to be relative when their maximum is
greater than 1, or an attribute \code{relative = TRUE} is attached to the
parameter.}

\item{lambda}{A vector of efficacy or control rates for each division part
specified by \code{divisions}, such that the probability of successful
control of an incursion when present at a part can be expressed via
\code{pr(control|presence) = 1 - exp(-lambda*allocation)}, for a given
continuous allocation of management control resources. Required unless
\code{unit_pr} is provided for discrete allocation.}

\item{unit_pr}{A vector of unit effectiveness for each division part
specified by \code{divisions}, such that the probability of successful
control of an incursion when present at a part can be expressed via
\code{pr(control|presence) = 1 - (1 - unit_pr)^allocation}, for a given
discrete allocation of management control resources. Required unless
\code{lambda} is provided for continuous allocation.}

\item{optimal}{The strategy used for finding an effective management
resource allocation. One of (maximum) \code{"saving"} (or cost-dependent
benefit), (maximum) \code{"benefit"} (independent of management resource
costs), (maximum) number of management \code{"successes"}, or (maximum)
overall system-wide \code{"effectiveness"} (probability of success), or
\code{"none"} for representing existing management resource allocation
designs only. Maximum \code{"effectiveness"} can only be used when actual
(not relative) establishment probability (\code{establish_pr}) values are
provided.}

\item{alloc_unit}{The descriptive unit to describe allocated management
resource quantities. One of \code{"units"}, \code{"hours"},
\code{"traps"}, \code{"treatments"}, \code{"removals"}, or user specified.}

\item{cost_unit}{The descriptive unit to describe allocated management
resource costs, cost-based benefit savings, and/or other costs when
applicable. One of \code{"$"}, \code{"hours"}, or user specified.}

\item{benefit}{A vector of values quantifying the benefit (or cost-based
saving) associated with allocated management resources at each division
part specified by \code{divisions}. Default is \code{NULL}. When the
benefit refers to cost-based savings (i.e. \code{optimal} is
\code{"saving"}), then the units should be consistent with the
\code{cost_unit} parameter.}

\item{alloc_cost}{A vector of cost per unit of allocated management
resources at each division part specified by \code{divisions}. Default is
\code{NULL}. Units should be consistent with the \code{cost_unit}
parameter.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs or time, at
each division part specified by \code{divisions}. Default is \code{NULL}.
Units should be consistent with \code{alloc_cost} when specified.
Otherwise the units should be consistent with the \code{alloc_unit}
parameter.}

\item{budget}{The cost budget or constraint for the resource allocation in
the management design. Default is \code{NULL}. Units should be consistent
with \code{alloc_cost} when specified. Otherwise the units should be
consistent with the \code{alloc_unit} parameter.}

\item{average_pr}{The desired (minimum) weighted average probability of
success (or effectiveness) of the management design (e.g. 0.95). The
weighted average is calculated using (relative) establishment probability
(\code{establish_pr}) values. Default is \code{NULL}.}

\item{overall_pr}{The desired (minimum) overall system-wide probability of
success (or effectiveness) of the management design (e.g. 0.95). Can only
be used when actual (not relative) establishment probability
(\code{establish_pr}) values are provided. Default is \code{NULL}.}

\item{min_alloc}{A vector of minimum permissible allocated management
resource quantities at each division part specified by \code{divisions}.
Used to avoid impractically low allocation quantities. Default is
\code{NULL}.}

\item{max_alloc}{A vector of maximum permissible allocated management
resource quantities at each division part specified by \code{divisions}.
Used to avoid impractically high allocation quantities. Default is
\code{NULL}.}

\item{discrete_alloc}{A logical to indicate that the allocated management
resource quantities at each division part specified by \code{divisions}
should be discrete integers. Used to allocate discrete management resource
units, such as traps or removals. Default is \code{FALSE} for continuous
resources quantities, such as control hours.}

\item{exist_alloc}{A vector of existing management resource quantities at
each division part specified by \code{divisions}. Should only be used to
represent existing management resource allocation designs when
\code{optimal = "none"}. Default is \code{NULL}.}

\item{exist_manage_pr}{A vector, or list of vectors, of probability of
success (or effectiveness) values for existing management resources at
each division part specified by \code{divisions}. Multiple existing
management success probabilities may be specified in a list. Default is
\code{NULL}.}

\item{previous_control}{A vector of control proportion values previously
applied to each division part specified by \code{divisions}, along with an
optional attached vector attribute \code{repeats} to indicate that the
control was applied repeatedly. These values modify the establishment
likelihood values utilised in the allocation design such that:
\code{mod_establish_pr = establish_pr*(1 - previous_control)^repeats}.
Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ControlDesign} class object (list) containing functions for
allocating resources, and calculating (unit and overall) likely management
success probabilities (or effectiveness):
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_divisions()}}{Get divisions object.}
\item{\code{get_dim_type()}}{Get dimension type.}
\item{\code{get_cost_unit()}}{Get cost unit.}
\item{\code{get_mod_establish_pr()}}{Get the modified establishment
likelihood (only when previous control is specified).}
\item{\code{get_allocation()}}{Get allocated resources via specified
strategy, utilizing savings, benefits, budget constraints, and/or
desired overall management success probability level.}
\item{\code{get_manage_pr()}}{Get the probability of success (or
effectiveness) values at each division part for the allocated
management design, combined with any existing success probabilities or
effectiveness specified via \code{exist_manage_pr}.}
\item{\code{get_average_pr()}}{Get the weighted average probability
of success (or effectiveness) of the management design. Weighted via
\code{establish_pr} values.}
\item{\code{get_overall_pr()}}{Get the overall system-wide probability
of success (or effectiveness) of the management design. Only available
when actual (not relative) \code{establish_pr} values are provided.}
\item{\code{save_design(...)}}{Save the management design as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{divisions} type, including the management
resource \code{allocation}, the corresponding probability of
management success (or effectiveness) values (\code{alloc_manage_pr})
when existing sensitivity is included, the overall probability of
management success values (\code{manage_pr}), \code{manage_cost}
(combined allocation and fixed costs), and a \code{summary} (CSV) of
the total allocation, costs (when applicable), the weighted average
probability of success (\code{average_pr}), and (when available) the
overall system-wide probability of success (or effectiveness) of the
management design (\code{overall_pr}), as well as the component of
each (\code{alloc_average_pr} and \code{alloc_overall_pr}
respectively) corresponding to the allocation when existing
sensitivity is included. \code{Terra} raster write options may be
passed to the function for saving grid-based designs.}
}
}
\description{
Builds a class to represent management control functionality for the
effective allocation of management control resources across one or more
divisions (invasion pathways, invasive species, spatial locations, etc.) via
Lagrange-based methods for optimizing objective functions specified with
management costs, savings, benefits, and/or overall probability of
management success (or effectiveness).
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}

Hauser, C. E., & McCarthy, M. A. (2009). Streamlining 'search and
destroy': cost-effective surveillance for invasive species management.
\emph{Ecology Letters}, 12(7), 683–692.
\doi{10.1111/j.1461-0248.2009.01323.x}

Giljohann, K. M., Hauser, C. E., Williams, N. S. G., & Moore, J. L.
(2011). Optimizing invasive species control across space: willow invasion
management in the Australian Alps. \emph{Journal of Applied Ecology},
48(5), 1286–1294. \doi{10.1111/j.1365-2664.2011.02016.x}

McCarthy, M. A., Thompson, C. J., Hauser, C., Burgman, M. A., Possingham,
H. P., Moir, M. L., Tiensin, T., & Gilbert, M. (2010). Resource allocation
for efficient environmental management. \emph{Ecology Letters}, 13(10),
1280–1289. \doi{10.1111/j.1461-0248.2010.01522.x}

Moore, A. L., McCarthy, M. A., & Lecomte, N. (2016). Optimizing ecological
survey effort over space and time.
\emph{Methods in Ecology and Evolution}, 7(8), 891–899.
\doi{10.1111/2041-210X.12564}
}
