% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageActions.R
\name{ManageActions}
\alias{ManageActions}
\alias{ManageActions.Region}
\title{Manage actions class builder}
\usage{
ManageActions(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = NULL,
  class = character(),
  ...
)

\method{ManageActions}{Region}(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{type}{One of \code{"detection"} (default), \code{"control"},
or \code{"removal"} to indicate the type of management actions applied.}

\item{stages}{Numeric vector of population stages (indices) to which
management actions are applied. Default is all stages (when set to
\code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps in which to apply
management actions. Default is all time steps (when set to \code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageActions} class object (list) containing a function for
applying simulated management actions:
\describe{
\item{\code{get_type()}}{Get the management actions type.}
\item{\code{get_label()}}{Get the management actions label used in
simulation results.}
\item{\code{get_stages()}}{Get the population stages to which management
actions are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which management actions are applied.}
\item{\code{apply(n, tm)}}{Apply management actions to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, providing the time
step \code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied actions.}
}
}
\description{
Builds a generic class for simulating the application of invasive species
management actions, such as detection, control, and removal.
}
