% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageContext.R
\name{ManageContext}
\alias{ManageContext}
\alias{ManageContext.default}
\title{Manage context class builder}
\usage{
ManageContext(
  species_names,
  species_types = c("pest", "weed", "disease"),
  impact_types = c("none", "ecological", "social", "economic"),
  action_types = c("none", "surveillance", "control", "removal"),
  manage_scope = c("optimal", "scenarios"),
  ...
)

\method{ManageContext}{default}(
  species_names,
  species_types = c("pest", "weed", "disease"),
  impact_types = c("none", "ecological", "social", "economic"),
  action_types = c("none", "surveillance", "control", "removal"),
  manage_scope = c("optimal", "scenarios"),
  ...
)
}
\arguments{
\item{species_names}{Vector of one or more invasive species (or genus)
names.}

\item{species_types}{Vector of types of invasive species. One or more of
\code{"pest"}, \code{"weed"}, or \code{"disease"}.}

\item{impact_types}{Vector of the types of impacts involved. Either
\code{"none"} when impacts are not considered, or one or more of
\code{"ecological"}, \code{"social"}, and/or \code{"economic"}.}

\item{action_types}{Vector of the types of management actions involved.
Either \code{"none"} when actions are not considered, or one or more of
\code{"surveillance"}, \code{"control"}, and/or \code{"removal"}.}

\item{manage_scope}{The scope of the resource management. One of
\code{"optimal"} resource allocation or management simulation
\code{"scenarios"}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageContext} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_species_names()}}{Get the invasive species names.}
\item{\code{get_species_types()}}{Get the types of invasive species:
"pest", "weed", or "disease".}
\item{\code{get_impact_types()}}{Get the impact types.}
\item{\code{get_action_types()}}{Get the management action types.}
\item{\code{get_manage_scope()}}{Get the management scope.}
}
}
\description{
Builds a class to represent the context of a biosecurity management resource
allocation or scenario analysis, including information about the invasive
species being managed, the of impacts considered, and the management
resources being utilised, plus the scope of the management specifying
whether to find optimal resource allocations across management responses,
invasive species, time, and/or space, or alternatively to analyse
simulations of management scenarios with different responses, or with and
without resources.
}
