% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageControls.R
\name{ManageControls}
\alias{ManageControls}
\alias{ManageControls.Region}
\title{Manage controls class builder}
\usage{
ManageControls(region, population_model, apply_stages = NULL, ...)

\method{ManageControls}{Region}(region, population_model, apply_stages = NULL, ...)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{apply_stages}{Numeric vector of population stages (indices) to which
management controls are applied. Default is all stages (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageControls} class object (list) containing a function
for accessing attributes and applying simulated management controls:
\describe{
\item{\code{get_type()}}{Get the type of management action ("control").}
\item{\code{apply(n)}}{Apply management controls to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, and return the
resulting population \code{n} along with attached attributes relating
to the newly applied controls.}
}
}
\description{
Builds a class for simulating the application of management controls, such
as suppressing growth, fertility, spread, or establishment.
}
