% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageControls.R
\name{ManageControls}
\alias{ManageControls}
\alias{ManageControls.Region}
\title{Manage controls class builder}
\usage{
ManageControls(region, population_model, stages = NULL, schedule = NULL, ...)

\method{ManageControls}{Region}(region, population_model, stages = NULL, schedule = NULL, ...)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{stages}{Numeric vector of population stages (indices) to which
management controls are applied. Default is all stages (when set to
\code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps in which to apply
management controls. Default is all time steps (when set to \code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageControls} class object (list) containing a function
for accessing attributes and applying simulated management controls:
\describe{
\item{\code{get_type()}}{Get the type of management action ("control").}
\item{\code{get_label()}}{Get the management actions label used in
simulation results (e.g. "control_growth").}
\item{\code{get_stages()}}{Get the population stages to which management
controls are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which management controls are applied.}
\item{\code{apply(n, tm)}}{Apply management controls to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, providing the time
step \code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied controls.}
}
}
\description{
Builds a class for simulating the application of invasive species management
controls, such as suppressing growth, fertility, spread, or establishment.
}
