% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageDesign.R
\name{ManageDesign}
\alias{ManageDesign}
\alias{ManageDesign.ManageContext}
\title{Manage design base class builder}
\usage{
ManageDesign(
  context,
  divisions,
  establish_pr = NULL,
  optimal = c("saving", "benefit", "effectiveness", "none"),
  alloc_unit = c("units", "hours", "traps", "treatments", "removals", "user"),
  cost_unit = c("$", "hours", "user"),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_manage_pr = NULL,
  class = character(),
  ...
)

\method{ManageDesign}{ManageContext}(
  context,
  divisions,
  establish_pr = NULL,
  optimal = c("saving", "benefit", "effectiveness", "none"),
  alloc_unit = c("units", "hours", "traps", "treatments", "removals", "user"),
  cost_unit = c("$", "hours", "user"),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  overall_pr = NULL,
  min_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_manage_pr = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{ManageContext} or inherited class object representing
the context of a biosecurity management resource allocation design.}

\item{divisions}{A \code{bsdesign::Divisions} or inherited class object
representing one or more divisions (management responses, strategies,
invasion pathways, invasive species, spatial locations, time, etc.) for
the management design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part (management
responses, strategies, invasion pathways, invasive species, spatial
locations, time, etc.) specified by \code{divisions}. Default is
\code{NULL}. Values are assumed to be relative when their maximum is
greater than 1, or an attribute \code{relative = TRUE} is attached to the
parameter.}

\item{optimal}{The strategy used for finding an effective management
resource allocation. One of (maximum) \code{"saving"} (or cost-dependent
benefit), (maximum) \code{"benefit"} (independent of management resource
costs), or (maximum) \code{"effectiveness"}, or \code{"none"} for
representing existing management resource allocation designs only.}

\item{alloc_unit}{The descriptive unit to describe allocated management
resource quantities. One of \code{"units"}, \code{"hours"},
\code{"traps"}, \code{"treatments"}, \code{"removals"}, or user specified.}

\item{cost_unit}{The descriptive unit to describe allocated management
resource costs, cost-based benefit savings, and/or fixed costs when
applicable. One of \code{"$"}, \code{"hours"}, or user specified.}

\item{benefit}{A vector of values quantifying the benefit (or cost-based
saving) associated with allocated management resources at each division
part (management responses, strategies, invasion pathways, invasive
species, spatial locations, time, etc.) specified by \code{divisions}.
Default is \code{NULL}. When the benefit refers to cost-based savings
(i.e. \code{optimal} is \code{"saving"}), then the units should be
consistent with the \code{cost_unit} parameter.}

\item{alloc_cost}{A vector of cost per unit of allocated management
resources at each division part (management responses, strategies,
invasion pathways, invasive species, spatial locations, time, etc.)
specified by \code{divisions}. Default is \code{NULL}. Units should be
consistent with the \code{cost_unit} parameter.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs or time, at
each division part (management responses, strategies, invasion pathways,
invasive species, spatial locations, time, etc.) specified by
\code{divisions}. Default is \code{NULL}. Units should be consistent with
\code{alloc_cost} when specified. Otherwise the units should be consistent
with the \code{alloc_unit} parameter.}

\item{budget}{The cost budget or constraint for the resource allocation in
the management design. Default is \code{NULL}. Units should be consistent
with \code{alloc_cost} when specified. Otherwise the units should be
consistent with the \code{alloc_unit} parameter.}

\item{overall_pr}{The desired (minimum) overall system-wide likely
management effectiveness or success rate of the management design
(e.g. 0.8) when \code{optimal} is \code{"effectiveness"}. Default is
\code{NULL}.}

\item{min_alloc}{A vector of minimum permissible allocated management
resource quantities at each division part (management responses,
strategies, invasion pathways, invasive species, spatial locations, time,
etc.) specified by \code{divisions}. Used to avoid impractically low
allocation quantities. Default is \code{NULL}.}

\item{discrete_alloc}{A logical to indicate that the allocated management
resource quantities at each division part (management responses,
strategies, invasion pathways, invasive species, spatial locations, time,
etc.) specified by \code{divisions} should be discrete integers. Used to
allocate discrete management resource units, such as traps or removals.
Default is \code{FALSE} for continuous resources quantities, such as
control hours.}

\item{exist_alloc}{A vector of existing management resource quantities at
each division part (management responses, strategies, invasion pathways,
invasive species, spatial locations, time, etc.) specified by
\code{divisions}. Should only be used to represent existing management
resource allocation designs when \code{optimal = "none"}. Default is
\code{NULL}.}

\item{exist_manage_pr}{A vector, or list of vectors, of existing management
effectiveness or success probability values for existing management
resources at each division part (management responses, strategies,
invasion pathways, invasive species, spatial locations, time, etc.)
specified by \code{divisions}. Multiple existing management success
probabilities may be specified in a list. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageDesign} class object (list) containing functions for
for allocating resources, and calculating (unit and overall) likely
management effectiveness or success rates:
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_divisions()}}{Get divisions object.}
\item{\code{get_allocation()}}{Get allocated resources via specified
strategy, utilizing savings, benefits, budget constraints, and/or
desired detection confidence level.}
\item{\code{get_manage_pr()}}{Get the division part management
effectiveness or success probabilities of the allocated management
design combined with any existing effectiveness or success
probabilities specified via \code{exist_manage_pr}.}
\item{\code{get_overall_pr()}}{Get the overall system-wide likely
management effectiveness or success rate of the management design.}
\item{\code{save_design(...)}}{Save the management design as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{divisions} type, including the management
resource \code{allocation}, the likely management effectiveness or
success rates (\code{manage_pr}), and a \code{summary} (CSV) of the
total allocation, total costs (when applicable), and the overall
system-wide effectiveness or likely management success rate.
\code{Terra} raster write options may be passed to the function for
saving grid-based designs.} ####
}
}
\description{
Builds a base class to represent management design functionality for the
effective allocation of management resources across one or more divisions
(management responses, strategies, invasion pathways, invasive species,
spatial locations, time, etc.) via methods that utilize management costs,
savings, benefits, and/or overall management effectiveness or likely success
rate.
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}
}
