% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageDivisions.R
\name{ManageDivisions}
\alias{ManageDivisions}
\alias{ManageDivisions.Raster}
\alias{ManageDivisions.SpatRaster}
\alias{ManageDivisions.matrix}
\alias{ManageDivisions.data.frame}
\title{Manage Divisions class builder}
\usage{
ManageDivisions(x, ...)

\method{ManageDivisions}{Raster}(x, ...)

\method{ManageDivisions}{SpatRaster}(x, ...)

\method{ManageDivisions}{matrix}(x, ...)

\method{ManageDivisions}{data.frame}(x, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
representing a grid-based spatial region (template). The spatial division
of locations may also be defined via a data frame (or matrix) of location
coordinates in longitude and latitude (WGS84) with explicitly named
columns "lon" and "lat". Alternatively, other none-spatial divisions, such
as categories, may be defined via a data frame (or matrix) with one or
more rows (one per part) but without coordinate columns.}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageDivisions} class object (list) containing functions
for accessing attributes, checking compatibility of objects with the
divisions, and to maintain and calculate spatial data:
\describe{
\item{\code{get_type()}}{Get the type of representation: spatial "grid"
(raster cells), "patch" (coordinates), or "other" (none-spatial).}
\item{\code{get_parts()}}{Get the number of parts (locations,
categories, management resource units, etc.) that are specified for
the design.}
\item{\code{is_compatible(y)}}{Check the compatibility of object
\code{y} with the divisions defined by \code{x}.}
\item{\code{get_template(empty = FALSE)}}{Get the spatial template when
the \code{type} is "grid", with either zeros in non-NA locations
(default), or with no values when \code{empty = TRUE}.}
\item{\code{get_rast(values)}}{Get spatial \code{terra::SpatRaster} set
with a single value or a vector of values matching the number of
locations when \code{type} is "grid".}
\item{\code{get_indices()}}{Get cell indices of grid or patch locations
that are included in the design when the \code{type} is "grid".}
\item{\code{get_res()}}{Get the spatial cell resolution (in m) of the
region when the \code{type} is "grid".}
\item{\code{is_included(indices)}}{Check if cell \code{indices} of grid
locations are included (non-NA cells) in the design when the
\code{type} is "grid", and return a logical vector indicating the
inclusion of each index.}
\item{\code{get_data()}}{Get the data frame specifying the division when
\code{type} is "patch" or "other".}
\item{\code{get_coords(extra_cols = FALSE)}}{Get a data frame of
location coordinates when \code{type} is "patch" or "grid", as well as
optional extra named columns from the original patch-based location
data.}
\item{\code{get_feat()}}{Get spatial \code{terra::SpatVector} features
or points when \code{type} is "grid" or "patch".}
}
}
\description{
Builds a class to represent one or more divisions (parts, locations,
categories, etc.) specified for a management resource allocation design.
It may be defined via a raster layer with active (non-NA) cells, via a data
frame of locations or patches with spatial coordinates, or via a data frame
without coordinates for none-spatial divisions, such as disease categories.
}
