% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageImpacts.R
\name{ManageImpacts}
\alias{ManageImpacts}
\title{Manage impacts class builder}
\usage{
ManageImpacts(impacts, population_model, impact_stages = NULL, ...)
}
\arguments{
\item{impacts}{A \code{bsimpact::ImpactAnalysis} or inherited class object
specifying how environment, social, and/or economic impacts are
calculated, classified, and/or combined based on asset values,
classifications, and/or incursion loss rates.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{impact_stages}{Numeric vector of population stages (indices) that
contribute to impacts. Default is all stages (when set to \code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageImpacts} class object (list) containing functions for
getting the impact context and performing impact calculations:
\describe{
\item{\code{get_context()}}{Get \code{bsimpact::Context} object.}
\item{\code{includes_combined()}}{Logical indicator for when impacts are
combined.}
\item{\code{calculate(n)}}{Perform impact calculations resulting from
incursion population vector or matrix \code{n}, and return a list
of impact values, including combined impacts when applicable.}
}
}
\description{
Builds a class for calculating incursion impacts during management
simulations. Encapsulates \code{bsimpact} impact analysis objects.
}
