% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageRemovals.R
\name{ManageRemovals}
\alias{ManageRemovals}
\alias{ManageRemovals.Region}
\title{Manage removals class builder}
\usage{
ManageRemovals(
  region,
  population_model,
  removal_pr = 1,
  remove_always = FALSE,
  detected_only = FALSE,
  radius = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)

\method{ManageRemovals}{Region}(
  region,
  population_model,
  removal_pr = 1,
  remove_always = FALSE,
  detected_only = FALSE,
  radius = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{bsspread::Region} or inherited class object
representing the spatial region (template) for the incursion management
simulations.}

\item{population_model}{A \code{bsspread::Population} or inherited class
object defining the population representation for the management
simulations.}

\item{removal_pr}{A single probability value (0-1), or vector of values for
each location specified by the \code{region}, to represent the likelihood
of removal at locations where the invasive species has been detected, as
indicated via a population attribute when present, else removal is applied
at all locations. Default is \code{1}, indicating all (detected)
occurrences are removed.}

\item{remove_always}{A logical indication of whether or not removal is
always to applied to locations where invasive species are present (even
when they are not detected by explicit surveillance). Default is
\code{FALSE}, indicating that removal is dependent on detection when
surveillance actions are present. Set to \code{TRUE} for locations where
some removal is likely regardless of explicit management efforts (e.g.
via the general public or property owners).}

\item{detected_only}{A logical indication of whether or not removal is only
applied to detected individuals (e.g. via traps). Default is \code{FALSE},
indicating that removal is applied to all individuals at locations where
the invasive species has been detected (e.g. via treatment).}

\item{radius}{Optional radius (m) of the removal. Stochastic removal is
applied to all locations within the specified radius of each location
where the invasive species has been detected. Default is \code{NULL},
indicating that removal is only applied at detected locations (when
specified via a population attribute).}

\item{stages}{Numeric vector of population stages (indices) to which
management removals are applied. Default is all stages (when set to
\code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps in which to apply
management removals Default is all time steps (when set to \code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ManageRemovals} class object (list) containing a function
for accessing attributes and applying simulated management removals:
\describe{
\item{\code{get_type()}}{Get the type of management action ("removal").}
\item{\code{get_label()}}{Get the management actions label used in
simulation results (i.e. "removed").}
\item{\code{get_stages()}}{Get the population stages to which management
removals are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which management removals are applied.}
\item{\code{apply(n, tm)}}{Apply management removals to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, providing the time
step \code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied removals.}
}
}
\description{
Builds a class for simulating the application of management removals or
eradication of an invasive species.
}
